Change Links, swap ItemsList and ItemsTable components
